refactor(vectorstore): split lancedb_store into focused modules
Behavior-preserving split of the 1264-line lancedb_store.rs into a
directory module: schema.rs (Arrow schema + batch conversion), filters.rs
(filter SQL translation/escaping), index.rs (ANN indexing config +
maintenance policy), worker.rs (command channel + worker loop), inner.rs
(async table operations), with mod.rs as the thin public facade.

Public API unchanged (LanceDbStore, IndexingConfig, filter_to_sql
re-exported from the same path). Added focused tests for newly exposed
pure logic: batch round-trip, dedupe last-wins, vector/dimension/schema
validation, SQL literal escaping, nested filter parenthesization, and
indexing-config rejection of non-integer/negative values.

Part of #22
